feat!: if_matches arguments changed, added docs
The `if_matches` macro takes a closure body instead of the entire
closure.

Added support for an optional match guard argument.

Added documentation comments for the `if_matches` macro.

BREAKING CHANGE: The last argument of the `if_matches` macro has been
changed from a closure that maps matched values to the body of the
closure.